Refactor BigQueryIO.Write helper transforms into separate files.
